+++ src/FastRouteRouter.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-fastroute for the canonical source repository
- * @copyright Copyright (c) 2015-2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2015-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-fastroute/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\Expressive\Router;
 
@@ -13,8 +15,29 @@ use FastRoute\RouteCollector;
 use FastRoute\RouteParser\Std as RouteParser;
 use Fig\Http\Message\RequestMethodInterface as RequestMethod;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use Zend\Expressive\Router\Exception;
 use Zend\Stdlib\ArrayUtils;
+
+use function array_key_exists;
+use function array_keys;
+use function array_merge;
+use function array_reduce;
+use function array_reverse;
+use function array_unique;
+use function dirname;
+use function file_exists;
+use function file_put_contents;
+use function implode;
+use function is_array;
+use function is_dir;
+use function is_string;
+use function is_writable;
+use function preg_match;
+use function restore_error_handler;
+use function set_error_handler;
+use function sprintf;
+use function var_export;
+
+use const E_WARNING;
 
 /**
  * Router implementation bridging nikic/fast-route.
@@ -24,7 +47,7 @@ class FastRouteRouter implements RouterInterface
     /**
      * Template used when generating the cache file.
      */
-    const CACHE_TEMPLATE = <<< 'EOT'
+    public const CACHE_TEMPLATE = <<< 'EOT'
 <?php
 return %s;
 EOT;
@@ -32,34 +55,17 @@ EOT;
     /**
      * @const string Configuration key used to enable/disable fastroute caching
      */
-    const CONFIG_CACHE_ENABLED = 'cache_enabled';
+    public const CONFIG_CACHE_ENABLED = 'cache_enabled';
 
     /**
      * @const string Configuration key used to set the cache file path
      */
-    const CONFIG_CACHE_FILE = 'cache_file';
-
-    /**
-     * HTTP methods that always match when no methods provided.
-     */
-    const HTTP_METHODS_EMPTY = [
-        RequestMethod::METHOD_GET,
-        RequestMethod::METHOD_HEAD,
-        RequestMethod::METHOD_OPTIONS,
-    ];
-
-    /**
-     * HTTP methods implicitly supported by any route
-     */
-    const HTTP_METHODS_IMPLICIT = [
-        RequestMethod::METHOD_HEAD,
-        RequestMethod::METHOD_OPTIONS,
-    ];
+    public const CONFIG_CACHE_FILE = 'cache_file';
 
     /**
      * Standard HTTP methods against which to test HEAD/OPTIONS requests.
      */
-    const HTTP_METHODS_STANDARD = [
+    public const HTTP_METHODS_STANDARD = [
         RequestMethod::METHOD_HEAD,
         RequestMethod::METHOD_GET,
         RequestMethod::METHOD_POST,
@@ -161,22 +167,20 @@ EOT;
     /**
      * Load configuration parameters
      *
-     * @param array $config Array of custom configuration options.
-     *
-     * @return void
+     * @param null|array $config Array of custom configuration options.
      */
-    private function loadConfig(array $config = null)
+    private function loadConfig(array $config = null) : void
     {
         if (null === $config) {
             return;
         }
 
         if (isset($config[self::CONFIG_CACHE_ENABLED])) {
-            $this->cacheEnabled = (bool)$config[self::CONFIG_CACHE_ENABLED];
+            $this->cacheEnabled = (bool) $config[self::CONFIG_CACHE_ENABLED];
         }
 
         if (isset($config[self::CONFIG_CACHE_FILE])) {
-            $this->cacheFile = (string)$config[self::CONFIG_CACHE_FILE];
+            $this->cacheFile = (string) $config[self::CONFIG_CACHE_FILE];
         }
 
         if ($this->cacheEnabled) {
@@ -190,20 +194,13 @@ EOT;
      * Uses the HTTP methods associated (creating sane defaults for an empty
      * list or Route::HTTP_METHOD_ANY) and the path, and uses the path as
      * the name (to allow later lookup of the middleware).
-     *
-     * @param Route $route
      */
-    public function addRoute(Route $route)
+    public function addRoute(Route $route) : void
     {
         $this->routesToInject[] = $route;
     }
 
-    /**
-     * @param  Request $request
-     *
-     * @return RouteResult
-     */
-    public function match(Request $request)
+    public function match(Request $request) : RouteResult
     {
         // Inject any pending routes
         $this->injectRoutes();
@@ -214,15 +211,6 @@ EOT;
         $method     = $request->getMethod();
         $dispatcher = $this->getDispatcher($dispatchData);
         $result     = $dispatcher->dispatch($method, $path);
-
-        if ($result[0] === Dispatcher::METHOD_NOT_ALLOWED
-            && in_array($method, self::HTTP_METHODS_IMPLICIT, true)
-        ) {
-            if (false !== ($result = $this->probeIntrospectionMethod($method, $path, $dispatcher))) {
-                return $this->marshalImplicitMethodResult($method, $result);
-            }
-            return $this->marshalFailedRoute($result);
-        }
 
         return $result[0] !== Dispatcher::FOUND
             ? $this->marshalFailedRoute($result)
@@ -244,16 +232,16 @@ EOT;
      *     in route used to generate URI.
      *
      * @return string URI path generated.
-     * @throws Exception\InvalidArgumentException if the route name is not known
+     * @throws Exception\RuntimeException if the route name is not known
      *     or a parameter value does not match its regex.
      */
-    public function generateUri($name, array $substitutions = [], array $options = [])
+    public function generateUri(string $name, array $substitutions = [], array $options = []) : string
     {
         // Inject any pending routes
         $this->injectRoutes();
 
         if (! array_key_exists($name, $this->routes)) {
-            throw new Exception\InvalidArgumentException(sprintf(
+            throw new Exception\RuntimeException(sprintf(
                 'Cannot generate URI for route "%s"; route not found',
                 $name
             ));
@@ -290,8 +278,8 @@ EOT;
                 }
 
                 // Check substitute value with regex
-                if (! preg_match('~^' . $part[1] . '$~', $substitutions[$part[0]])) {
-                    throw new Exception\InvalidArgumentException(sprintf(
+                if (! preg_match('~^' . $part[1] . '$~', (string) $substitutions[$part[0]])) {
+                    throw new Exception\RuntimeException(sprintf(
                         'Parameter value for [%s] did not match the regex `%s`',
                         $part[0],
                         $part[1]
@@ -307,7 +295,7 @@ EOT;
         }
 
         // No valid route was found: list minimal required parameters
-        throw new Exception\InvalidArgumentException(sprintf(
+        throw new Exception\RuntimeException(sprintf(
             'Route `%s` expects at least parameter values for [%s], but received [%s]',
             $name,
             implode(',', $missingParameters),
@@ -318,13 +306,10 @@ EOT;
     /**
      * Checks for any missing route parameters
      *
-     * @param array $parts
-     * @param array $substitutions
-     *
      * @return array with minimum required parameters if any are missing or
      *     an empty array if none are missing
      */
-    private function missingParameters(array $parts, array $substitutions)
+    private function missingParameters(array $parts, array $substitutions) : array
     {
         $missingParameters = [];
 
@@ -352,10 +337,8 @@ EOT;
 
     /**
      * Create a default FastRoute Collector instance
-     *
-     * @return RouteCollector
      */
-    private function createRouter()
+    private function createRouter() : RouteCollector
     {
         return new RouteCollector(new RouteParser, new RouteGenerator);
     }
@@ -371,7 +354,7 @@ EOT;
      *
      * @return Dispatcher
      */
-    private function getDispatcher($data)
+    private function getDispatcher($data) : Dispatcher
     {
         if (! $this->dispatcherCallback) {
             $this->dispatcherCallback = $this->createDispatcherCallback();
@@ -384,10 +367,8 @@ EOT;
 
     /**
      * Return a default implementation of a callback that can return a Dispatcher.
-     *
-     * @return callable
      */
-    private function createDispatcherCallback()
+    private function createDispatcherCallback() : callable
     {
         return function ($data) {
             return new Dispatcher($data);
@@ -399,29 +380,20 @@ EOT;
      *
      * If the failure was due to the HTTP method, passes the allowed HTTP
      * methods to the factory.
-     *
-     * @param array $result
-     *
-     * @return RouteResult
      */
-    private function marshalFailedRoute(array $result)
+    private function marshalFailedRoute(array $result) : RouteResult
     {
         if ($result[0] === Dispatcher::METHOD_NOT_ALLOWED) {
             return RouteResult::fromRouteFailure($result[1]);
         }
 
-        return RouteResult::fromRouteFailure();
+        return RouteResult::fromRouteFailure(Route::HTTP_METHOD_ANY);
     }
 
     /**
      * Marshals a route result based on the results of matching and the current HTTP method.
-     *
-     * @param array $result
-     * @param string $method
-     *
-     * @return RouteResult
      */
-    private function marshalMatchedRoute(array $result, $method)
+    private function marshalMatchedRoute(array $result, string $method) : RouteResult
     {
         $path  = $result[1];
         $route = array_reduce($this->routes, function ($matched, $route) use ($path, $method) {
@@ -441,17 +413,23 @@ EOT;
         }, false);
 
         if (false === $route) {
-            // This likely should never occur, but is present for completeness.
-            return RouteResult::fromRouteFailure();
+            return $this->marshalMethodNotAllowedResult($result);
         }
 
-        return RouteResult::fromRoute($route, $this->marshalMatchedParams($route, $result));
+        $params = $result[2];
+
+        $options = $route->getOptions();
+        if (! empty($options['defaults'])) {
+            $params = array_merge($options['defaults'], $params);
+        }
+
+        return RouteResult::fromRoute($route, $params);
     }
 
     /**
      * Inject queued Route instances into the underlying router.
      */
-    private function injectRoutes()
+    private function injectRoutes() : void
     {
         foreach ($this->routesToInject as $index => $route) {
             $this->injectRoute($route);
@@ -461,10 +439,8 @@ EOT;
 
     /**
      * Inject a Route instance into the underlying router.
-     *
-     * @param Route $route
      */
-    private function injectRoute(Route $route)
+    private function injectRoute(Route $route) : void
     {
         // Filling the routes' hash-map is required by the `generateUri` method
         $this->routes[$route->getName()] = $route;
@@ -480,10 +456,6 @@ EOT;
             $methods = self::HTTP_METHODS_STANDARD;
         }
 
-        if (empty($methods)) {
-            $methods = self::HTTP_METHODS_EMPTY;
-        }
-
         $this->router->addRoute($methods, $route->getPath(), $route->getPath());
     }
 
@@ -492,16 +464,14 @@ EOT;
      * FastRoute data generator.
      *
      * If caching is enabled, store the freshly generated data to file.
-     *
-     * @return array
      */
-    private function getDispatchData()
+    private function getDispatchData() : array
     {
         if ($this->hasCache) {
             return $this->dispatchData;
         }
 
-        $dispatchData = $this->router->getData();
+        $dispatchData = (array) $this->router->getData();
 
         if ($this->cacheEnabled) {
             $this->cacheDispatchData($dispatchData);
@@ -513,11 +483,10 @@ EOT;
     /**
      * Load dispatch data from cache
      *
-     * @return void
      * @throws Exception\InvalidCacheException If the cache file contains
      *     invalid data
      */
-    private function loadDispatchData()
+    private function loadDispatchData() : void
     {
         set_error_handler(function () {
         }, E_WARNING); // suppress php warnings
@@ -543,13 +512,13 @@ EOT;
     /**
      * Save dispatch data to cache
      *
-     * @param array $dispatchData
-     *
      * @return int|false bytes written to file or false if error
      * @throws Exception\InvalidCacheDirectoryException If the cache directory
      *     does not exist.
      * @throws Exception\InvalidCacheDirectoryException If the cache directory
      *     is not writable.
+     * @throws Exception\InvalidCacheException If the cache file exists but is
+     *     not writable
      */
     private function cacheDispatchData(array $dispatchData)
     {
@@ -569,80 +538,22 @@ EOT;
             ));
         }
 
+        if (file_exists($this->cacheFile) && ! is_writable($this->cacheFile)) {
+            throw new Exception\InvalidCacheException(sprintf(
+                'The cache file %s is not writable',
+                $this->cacheFile
+            ));
+        }
+
         return file_put_contents(
             $this->cacheFile,
             sprintf(self::CACHE_TEMPLATE, var_export($dispatchData, true))
         );
     }
 
-    /**
-     * Dispatch the given path against the set of standard methods to see if a
-     * match exists.
-     *
-     * Call this method for failed HEAD or OPTIONS requests, to see if another
-     * method matches; if so, return the match.
-     *
-     * @param string $method
-     * @param string $path
-     * @param Dispatcher $dispatcher
-     *
-     * @return false|array False if no match found, array representing the match
-     *     otherwise.
-     */
-    private function probeIntrospectionMethod($method, $path, Dispatcher $dispatcher)
-    {
-        foreach (self::HTTP_METHODS_STANDARD as $testMethod) {
-            if ($method === $testMethod) {
-                continue;
-            }
-            $result = $dispatcher->dispatch($testMethod, $path);
-            if ($result[0] === Dispatcher::FOUND) {
-                return $result;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @param string $method
-     * @return RouteResult
-     */
-    private function marshalIMplicitMethodResult($method, array $result)
+    private function marshalMethodNotAllowedResult(array $result) : RouteResult
     {
         $path = $result[1];
-        $route = array_reduce($this->routes, function ($matched, $route) use ($method, $path) {
-            if ($matched) {
-                return $matched;
-            }
-
-            if ($path !== $route->getPath()) {
-                return $matched;
-            }
-
-            return $route;
-        }, false);
-
-        $implicitRoute = new Route(
-            $path,
-            $route->getMiddleware(),
-            $this->marshalAllowedMethodsForPath($path),
-            $route->getName()
-        );
-        $implicitRoute->setOptions($route->getOptions());
-
-        return RouteResult::fromRoute(
-            $implicitRoute,
-            $this->marshalMatchedParams($implicitRoute, $result)
-        );
-    }
-
-    /**
-     * @param string $path
-     * @return array
-     */
-    private function marshalAllowedMethodsForPath($path)
-    {
         $allowedMethods = array_reduce($this->routes, function ($allowedMethods, $route) use ($path) {
             if ($path !== $route->getPath()) {
                 return $allowedMethods;
@@ -651,16 +562,8 @@ EOT;
             return array_merge($allowedMethods, $route->getAllowedMethods());
         }, []);
 
-        return array_unique($allowedMethods);
-    }
+        $allowedMethods = array_unique($allowedMethods);
 
-    private function marshalMatchedParams(Route $route, array $result)
-    {
-        $params = $result[2];
-        $options = $route->getOptions();
-        if (! empty($options['defaults'])) {
-            $params = array_merge($options['defaults'], $params);
-        }
-        return $params;
+        return RouteResult::fromRouteFailure($allowedMethods);
     }
 }